test de cuenta corriente y caja de ahorro

+++ PROG2-2025-EJ1-Zehnder/src/main/java/org/Practica/Cuenta.java
@@ -17,7 +17,7 @@ public abstract class Cuenta implements IGestionSaldo {
     }
 
     @Override
-    public synchronized double getOperaciones() {
+    public synchronized int getOperaciones() {
         synchronized (this) {
             return operaciones;
         }
