+++ PROG2-2025-EJ1-Zehnder/src/main/java/org/Practica/IGestionSaldo.java
@@ -5,6 +5,6 @@ public interface IGestionSaldo {
     boolean agregarSaldo(double monto);
     boolean quitarSaldo(double monto);
     double getSaldo();
-    double getOperaciones();
+    int getOperaciones();
 
 }
